refactor(server_events): reorganize event handling and rename slot_migrations module.
